requirements.txt
Flask
Flask-Login
python-dotenv
pytz
gunicorn

Flask-Migrate
Flask-SQLAlchemy
PyMySQL

pytest
pytest-cov
pytest-flask

PyJWT
cryptography
requests
